+++ packages/web-components/@sfx/product/stories/index.ts
@@ -1,11 +1,11 @@
 import { storiesOf } from '@storybook/html';
 import { withKnobs, text } from '@storybook/addon-knobs';
-import { ProductModel } from '@sfx/product';
+import { Product } from '@sfx/events';
 
 import { getDisplayCode, getProducts } from '../../../../../.storybook/common';
 import '../src/index';
 
-function getProductComponent(product: ProductModel = {}, codeSnippet: boolean = false): string {
+function getProductComponent(product: Product = {}, codeSnippet: boolean = false): string {
   const productInfo = text('Product Info', JSON.stringify(product));
 
   return codeSnippet === false
@@ -79,7 +79,7 @@ storiesOf('Components|Product', module)
 
         ### The SF-X Product component renders with product data populated via the \`product\` attribute.
         * ***Disclaimer***: although possible, it is not recommended to pass data objects via an attribute.
-          * Refer to the \`ProductModel\` and \`ProductVariantsModel\` interfaces for the accepted data format.
+          * Refer to the \`Product\` and \`ProductVariants\` interfaces for the accepted data format.
           * To modify the data within the \`product\` attribute in this story:
             1. Visit the **Knobs** tab and modify the data inside the "Product Data" field.
             2. Observe that the component is updated with the new data.