+++ packages/web-components/@sfx/search-box/src/events.ts
@@ -1,7 +1,0 @@
-export const SEARCHBOX_EVENT = {
-  UPDATE_SEARCH_TERM: 'sfx::update_search_term',
-  SEARCH_REQUEST: 'sfx::search_request',
-  SEARCHBOX_CLEAR_CLICK: 'sfx::searchbox_clear_click',
-  SEARCHBOX_CLICK: 'sfx::searchbox_click_event',
-  SEARCHBOX_CHANGE: 'sfx::searchbox_change'
-};