+++ packages/web-components/@sfx/sayt/CHANGELOG.md
@@ -9,3 +9,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - SFX-191: Added the `sayt` component.
   - This component renders `sayt` related components such as `sfx-autocomplete` and `sfx-products`.
   - SFX-333: It emits events for requesting products and autocomplete terms.
+  - SFX-248: The `@sfx/events` package is used for event names and payload interfaces.